nics: list dpdk ports
dpdk ports can't be seen by kernel tools.
We wish to know which dpdk ports are in the system in order to
bind them to ovs-dpdk bridge.
We query 'lshw -json' in order to get a json formatted response
of system devices, take only network class devices which have a
'handle' key and don't have logical name - meaning they are physical
network devices which are not recognized by the kernel.

Also, added fake network info for dpdk ports.

Change-Id: Ia07f5a1052146c0eea6f0e19aa22d68ff4148646
Signed-off-by: Irit Goihman <igoihman@redhat.com>