net: Expose an existing bug in network restoration
This patch exposes a bug that exists in the network restoration flow
but was missed due to a missing step in one test and a missing
assertion.

testVolatileConfig test now simulates a reboot by deleting the running
config (which resides under /var/run/).
In addition, the assertNetworkDoesntExist assertion should also check
against the libvirt DB, assuring no leftovers are left there.

In the next patch this scenario will be solved, passing the failing
testVolatileConfig functional test.

Change-Id: Ifad0c04df9a36fba43ee678bbec4f3500b6a98ec
Signed-off-by: Edward Haas <edwardh@redhat.com>