net: Modify restore-net to remove nets using setupNetworks
Overview:
The network restoration flow has a step in which acquired devices which
have been 'taken' by VDSM are restored.
When a device is acquired by VDSM, it is backed up, allowing later
rollback. The backed up data is cleared on setSafe or in case a
restoration is performed.

Problem:
When restoring in this step, an existing network (or partial network) is
removed by the underlying ifcfg configurator, but not through the proper
API.

Therefore, in such a scenario:
Clean host -> setupNetworks(add A) -> restore-net
(note that setSate() is missing)
Network A will be removed using ifcfg backup restoration flow without
issuing a setupNetworks(remove A).

One outcome of the missing proper removal of the network is that its
leftovers remain in the libvirt db.

This patch changes the restore flow such that it first removes extra
networks/bonds (which do not appear in the persisted config, but do
appear on the host) and only then attempts to restore the acquired
devices.

Change-Id: I387d5e3fd9be74e2789904ca8eaf9d867c9d4c3d
Signed-off-by: Edward Haas <edwardh@redhat.com>