virt: network: factor out update helper code
The Vm.setLinkAndNetwork() method is the only client
of vmxml.Element.replaceChild(), which we want to remove.

To do that, we start adding one test to cover the existing behaviour.
In turn to make this easier, we factor out the surrounding chunk of the
Vm.setLinkAndNetwork() code.

While this change is self sufficient and actually improves testability,
it is by no means a clear improvement or smart. We intentionally do
a dumb cut, because to properly clean this flow we will need a much
deeper refactoring, which is out of scope now.

We improve testability and modularity (the code is no worse than
before), so this change still have some merit.

Change-Id: I66e9618dbeb4280aede57e26ca389cea6fe1af31
Signed-off-by: Francesco Romani <fromani@redhat.com>