virt: Run job completed callback only for migrations
We introduced onJobCompleted callback to handle finished migration statistics
and other stuff.  However we can't be sure that the callback is run only for
remote host migration jobs.  We add a check for that.

We use SourceThread.started flag for that purpose.  Once a migration starts,
the flag is set.  If the migration finishes successfully, we keep the flag set
and we know that the callback must be from a migration (assuming it's not a
delayed callback from another domain job and that we don't run other domain
jobs in parallel with migrations).  If the migration fails and the VM continues
to operate then the flag is unset (this is change in handling of the flag but
it doesn't influence its original use), in which case we disable the callback
processing.

We've got a similar problem in _recover_status, but we'll handle that
separately, since we need more information from libvirt developers there and we
needn't backport _recover_status changes to 4.1.

Change-Id: I066906f4898d188ceac50101df6bf20adfa0ff89
Backport-To: 4.1
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>