tests: uncover test_interface_update is broken
We have a silently broken test. device_tests.test_interface_update is
broken because the success condition was too lax.

The test was deemed succesfull if the update XML is *sent* to the
fake libvirt domain, but this is just part of the flow.
The test actually fails just later when attempting to do port mirroring.
We do not have port mirroring, so the method is supposed to exit
quickly; but it fails right at the beginning, while checking what to do.

Since this test is checking the whole flow, and not just the XML update,
this means the test is broken.

This patch adds:
1. one check uncovering the real status of the test
2. the @brokentest decorator to signal this without breaking CI.

Change-Id: I077f64705b8d3b43c06a9a8948e43631701c6075
Signed-off-by: Francesco Romani <fromani@redhat.com>