tests: avoid to return lambda
The MockedProxy (which looks more like a fake) in device_test
may return a lambda. This will lead to a ETree serialization error,
should the XML be needed in the test code.

This doesn't actually trigger because we never need to recover
the device state, but it is wrong anyway, so this patch removes
the lambda to avoid future headaches.

Change-Id: I7bd8c77a49b5c3a8de9ccd20d48cafc678d2cb91
Signed-off-by: Francesco Romani <fromani@redhat.com>