caps: silence pylint distutils errors
pylint throws these errors:
E:218, 4: No name 'version' in module 'distutils' (no-name-in-module)
E:218, 4: Unable to import 'distutils.version' (import-error)

It happens because distutils version is created dynamically
and not being recognized by pylint.
It's also a good practice to import modules and not member
from modules so the import statement now imports version
module and not a member from 'version'.

Change-Id: If1e07243015801e4ba3020946e17e823a257df22
Signed-off-by: Irit Goihman <igoihman@redhat.com>