qcow2: Calculate estimated size for converted qcow volumes.
Adding a utility module for estimating qcow2 file size once
converted from raw to qcow2.
The estimation takes in consider also the qcow header size
based on the implementation being done in qemu block/qcow2.c

The following is a test being done on a 10GB QCOW volume
with compatibility of 1.1.
The actual size of the disk after executing qemu-convert was 1.00036621
GB, while the estimated size by the new function was 1.00177 GB

Change-Id: Ia87df2b0d5378f93c5cb2cc68a37458fe3b4467b
Bug-Url: https://bugzilla.redhat.com/1358717
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>