gluster: avoid over allocation in lvs
While provisioning gluster bricks, 16GB or 0.005% is reserved
for thin pool meta data. But lvs is created with actual disk size.
It results in over allocation of LV than what is available in
thinpool. This configuration can lead to storage corruption when
thinpool is fully utilized and xfs thinks still there is a space in lv.

Fixing this issue by avoiding over allocation. Now thin lv
will be created with the size same as thin pool without
over commiting.

Change-Id: I421401246b77c33e3ad02d08b3c0c06335c71446
Bug-Url: https://bugzilla.redhat.com/1412455
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>