virt: Limit migration recovery status to previously migrating VMs
We introduced a check for running migrations in recovery.  We utilize
migration.ongoing function for the check.

However migration.ongoing assumes that the domain has already been migrating to
another host, it doesn't work correctly in the general case, when we don't know
anything about the VM migration status in advance.  The migration.ongoing check
is based on checking the presence of a domain job since there is no much better
way to check for ongoing migrations in libvirt.  However the domain jobs are
not limited to host-to-host migrations, they can perform other tasks such as
hibernation or making a snapshot.

To satisfy the assumption of migration.ongoing we check for MIGRATION_SOURCE
status in recovery.

Change-Id: I37ce91f9490206fe2c68504dd17aa4326764d6ed
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>