xmlrpc: stop detection
Xmlrpc detector was responsible to detect http post which was used by
xmlrpc and get and put methods used by image upload and download logic.
We do not need to detect post anymore but we still need to detect put
and get methods.

We want to deal in the same way for all not supported http verbs and the
default behavior is to drop connection.


Change-Id: I6ab7d42efab98f23736d323b2b0b614a568b7112
Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>