xmlrpc: stop using during migration
All supported versions of vdsms use jsonrpc so there is no need to
fallback to xmlrpc anymore. When we are unable to use jsonrpc we fail
and let the issue propagate to the caller.

We want to keep the logic without the changes so it means that we need
cannonize_host_port which used to be part of vdscli.


Change-Id: I4c0301a8e4cd020e8f59f665da718394603a875c
Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>