tests: Add fake produceVolume
In patch I94ea20 (Cold Merge: Shrink base volume), we are going
to invoke shrinkToOptimalSize. This method is defined on the volume
class and not on the volume manifest. In this patch we define a fake
volume class that defines a shrinkToOptimalSize method that does nothing but
pass. Finally, we define a produceVolume on FakeSD class that returns the
FakeVolume.

Change-Id: I9d4729e9ff559c14f7b32bfc63699ce017ee3fa1
Bug-Utl: https://bugzilla.redhat.com/1420405
Signed-off-by: Ala Hino <ahino@redhat.com>