virt: devices: start the from_xml_tree method
Let's start adding to devices the capability to be initialized
from their XML representation.
The end goal is to make possible to properly initialize a device
from either the usual parameters dictionary or a valid XML snippet.

This is only about device construction, no changes about the
state management whatsoever.

Change-Id: I21e80625f1b86cd5a49571d2e121ee479535d8d3
Signed-off-by: Francesco Romani <fromani@redhat.com>