sp: Mark helper method as @unsecured
In commit 7cf19dafd7cd (storage_mailbox: make inbox/outbox mailbox
args), we added a helper that is used before the spm is started, but the
helper was not marked as @unsecure. This cause the call to fail with:

    File "/usr/share/vdsm/storage/sp.py", line 485, in __createMailboxMonitor
        outbox = self._master_volume_path("inbox")
      File "/usr/lib/python2.7/site-packages/vdsm/storage/securable.py",
    line 77, in wrapper
        raise SecureError("Secured object is not in safe state")
    SecureError: Secured object is not in safe state

As this helper doesn't change the state of the storage pool, there is no
reason to treat it as a secured method, which is the default for this
class.

Change-Id: Icf92b9474c9000840a5c15e3b91f2ced4d02aca2
Signed-off-by: Nir Soffer <nsoffer@redhat.com>