sp: Mark _finalizePoolUpgradeIfNeeded as @unsecure
StoragePool._finalizePoolUpgradeIfNeeded while a host is not the SPM. In
this case it will fail with this error:

    Traceback (most recent call last):
      File "/usr/lib/python2.7/site-packages/vdsm/utils.py", line 372, in wrapper
	return f(*a, **kw)
      File "/usr/lib/python2.7/site-packages/vdsm/concurrent.py", line 177, in run
	return func(*args, **kwargs)
      File "/usr/lib/python2.7/site-packages/vdsm/storage/securable.py", line 78, in wrapper
	return method(self, *args, **kwargs)
      File "/usr/share/vdsm/storage/sp.py", line 207, in _upgradePoolDomain
	self._finalizePoolUpgradeIfNeeded()
      File "/usr/lib/python2.7/site-packages/vdsm/storage/securable.py", line 76, in wrapper
	raise SecureError("Secured object is not in safe state")
    SecureError: Secured object is not in safe state

When this fails, we do not unregister StoragePool._upgradeCallback from domain
state changes events. Each time a domain changes state to valid, we will try to
to perform a useless domain upgrade, creating more noise in the log.

I think this issue was introduced by commit 4677d5b2dbcf (sp:
deactivateSd - remove domain from pending for upgrade list). Not
sure why the issue surfaced only recently.

Change-Id: I39c65223dd1c03928e4d74efa11c2177f8f9b854
Bug-Url: https://bugzilla.redhat.com/1400534
Signed-off-by: Nir Soffer <nsoffer@redhat.com>