net tests: Add grace/synchronization period for mode4 bonds.
Although mode 4 bonds may be reported up, they may still not be fully
synchronized with their partner bond. The rate in which dedicated
packets are sent to the slaves is set/determined via the "lp_interval"
option -- wait the interval and some predetermined grace period to allow
the bonds to become synchronized with their partner bonds.

Change-Id: I70ad2e8fbf2b54971eba79f8de44e199e54ae4e3
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>