mailbox: Decrease the chance of checksum failures
The HSM mailbox is writing 4096 bytes to the SPM inbox. The last 4 bytes
of this data are the checksum of the first 4092 bytes. We used to write
this data using block size of 512 bytes, doing 8 writes. From the time
the first write is complete, until the last write is complete, this host
mailbox is in invalid state; if the SPM reads the mailbox, the checksum
verification will fail, and the SPM will ignore the messages from this
host.

There is no way to write 4096 bytes atomically, so we cannot fix this
issue. But we can decrease the chance of the failure by doing one write
of 4096 bytes, instead of 8 writes of 512 bytes. There is still chance
of failure since writing on the remote storage is probably done using
512 blocks, but the chance of reading incomplete write is much lower.

Since both reading and writing was done using 512 bytes blocks, both
were modified to use 4096 block size. The SPM mailbox is already writing
replies using 4096 block size.

Change-Id: I3e170d69437a9fb86bac7b65d6fcd48971375e99
Bug-Url: https://bugzilla.redhat.com/1426762
Signed-off-by: Nir Soffer <nsoffer@redhat.com>