tests: Make test logs easier to understand
Previously we create the inbox and outbox with random names, which makes
it harder to understand the logs when tests fail. Replace the
mailbox_file() helper with standard make_env() creating a temporary
directory with "inbox" and "outbox" files.

Here are example logs with the new setup:

    DEBUG   [storage.Misc.excCmd] (mailbox-spm) /usr/bin/taskset --cpu-list
    0-3 dd if=/var/tmp/tmpGVQ6NU/inbox iflag=direct,fullblock count=1
    bs=40960 (cwd None)

    DEBUG   [storage.Misc.excCmd] (mailbox-hsm) /usr/bin/taskset --cpu-list
    0-3 /usr/bin/dd if=/var/tmp/tmpGVQ6NU/outbox iflag=direct,fullblock
    bs=512 count=8 skip=56 (cwd None)

    DEBUG   [storage.Misc.excCmd] (mailbox-hsm) /usr/bin/taskset
    --cpu-list 0-3 /usr/bin/dd of=/var/tmp/tmpGVQ6NU/inbox iflag=fullblock
    oflag=direct conv=notrunc bs=512 seek=56 (cwd None)

It is clear now that the mailbox-spm thread is reading from the inbox,
and the mailabox-hsm thread is reading from the outbox and writing to
the inbox.

Change-Id: I51379c3df6452832f5865270ef05f91c606126c8
Signed-off-by: Nir Soffer <nsoffer@redhat.com>