tests: Add helpers for running hsm and spm mailboxes
Creating SPM and HSM mailboxes is very verbose, and both of them create
a thread pool so we must stop them and wait until the thread pool is
destroyed. Add make_hsm_mailbox() and make_spm_mailbox() context
managers to make it easier to write tests.

The thread leak test cannot use the helpers because of the evil things
it does during the test. This tests will be probably be removed soon.

Change-Id: I5f750a814ea11c6cbfb12d0d1ae969a076d4b613
Signed-off-by: Nir Soffer <nsoffer@redhat.com>