virt: host: stats: do not replace HostMonitor
Some time ago we moved the HostMonitor operation inside the periodic
framework. This was made out of convenience, to kill one extra thread,
under the assumption that HostMonitor never blocks.

But from https://bugzilla.redhat.com/show_bug.cgi?id=1419856#c15
we learn that indeed HostMonitor can indeed block under rare, still
not completely clear, circumstances.

If more than two thread enter the netlink interface, all of them block.
There is perhaps a bug in libnl, but we should never access concurrently
this interface in the first place. So, if one HostMonitor is slow, we
must never replace it, so we can never access the same interface more
than once.

This is a partial fix. It is still unclear why the first HostMonitor is
slow, so why we end up with a timeout in the first place; but still,
given
1. the original intentions for moving HostMonitor in the periodic
framework
2. the fact that we don't need the host stats in the critical flow
3. the fact that we should never access concurrently the netlink
interface

then it is safe to never replace the HostMonitor operation. If it is
blocked, it will stay stuck unless it unblocks, or Vdsm is restarted.

Change-Id: I60ccd4e0e239ce5dfa2c90947bd6cd59a23e51b3
Related-To: https://bugzilla.redhat.com/1419856
Backport-To: 4.1
Backport-To: 4.0
Signed-off-by: Francesco Romani <fromani@redhat.com>