vm: tests: add test for initial vcpu limit setup
We update the vm vcpu limit both on startup - if SLA QOS policy is given
and after policy updates received from clients (e.g. Engine).
We had tests for the latter, not for the former cases.

This patch adds the missing test.
For the sake of the practicality, we take one shortcut and we test
a private Vm method. This is bad, but not having tests it is even
worse, as the related BZ demonstrates.

Change-Id: I7ae776e4f2abee2d4716ec260d00954fdd0b4eac
Related-To: https://bugzilla.redhat.com/1426727
Backport-To: 4.1
Signed-off-by: Francesco Romani <fromani@redhat.com>