storage: Do not make volume ILLEGAL before sparsifying it
Since virt-sparsify tool cannot leave the volume in corrupted state,
there is no need to mark the volume as ILLEGAL before running the tool.
Otherwise, if the storage domain becomes unavailable during the
operation, the volume will be left in ILLEGAL state and its futher usage
will be impossible. This patch helps to avoid such a scenario.

This patch introduces VolumeInfo class that is used by sparsify_volume
job instead of CopyDataDivEndpoint class that doesn't provide the
functionality needed. The VolumeInfo class is made common for
sparsify_volume and amend_volume verbs.

Change-Id: I8a7fbac2f5fe1bb62bb88f67038fedb3b815df3f
Bug-Url: https://bugzilla.redhat.com/1425161
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>