virt: Run job completed actions after recovered migrations
VM.onJobCompleted checks self._migrationSourceThread.started flag to know
whether the callback was invoked after a migration.  However if Vdsm was
restarted during the migration, the flag is not set since the source thread is
not actually running.  So in such a case onJobCompleted returns immediately
without performing the post-migration actions.

We fix it by adding a check for a recovered migration at the place.  We must
also ensure that SourceThread._recovery is reset after unsuccessful migration,
otherwise the post-migration actions would be called for all subsequent domain
jobs.  This doesn't work in all situations since libvirt may not notify us
after failed migrations so SourceThread._recovery gets reset only when the
migration is explicitly canceled.

Change-Id: I2eaacbfc48bee4863ec75b29e120fc5aa7b31d10
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>