hostdev: don't process devices that are gone
Processing the device tree is inherently racy - the initial list of
devices could've changed, but this is not really reflected in our
object.

(Un)luckily, methods on virNodeDev underlying struct are passed to the
libvirt. That means if the device disappeared, the call will raise
libvirtError which we can catch and skip the device.

Change-Id: If73429d383a39061bb21ff11d08456998971edca
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1428692
Signed-off-by: Martin Polednik <mpolednik@redhat.com>