storage.asyncutils: Introduce the asyncutils module
This module provides the LoopingCall class, implementing the common
pattern of running a callback every interval seconds.

Example usage:

    lc = LoopingCall(loop, callback, arg1, arg2)
    lc.start(10)

The first call is invoked when calling start; then the callback will be
invoke every 10 seconds without the need to reschedule the next call on
each callback.

To delay the first call, use EventLoop.call_after():

    loop.call_after(10, lc.start, 10)

To stop the calls:

    lc.stop()

LoopingCall adds two important warnings:

- Delayed call: if a call miss the expected deadline by more than 0.5
  seconds (configurable), log a warning about the delay.

- Missed deadlines: if the next deadline was missed when scheduling the
  next call, log a warning about the missed deadlines.

Both warnings mean the event loop was too busy or blocked by a bad
callback; both cases are not expected and we want to detect them easily
in the field.

LoopingCall will replace similar code in the storage check module,
allowing better logging when a checker is blocked on storage. This may
also be useful for the new block jobs monitor.

Change-Id: Ia8d91de2f2763455985010a42c876c7f7e1c24a6
Bug-Url: https://bugzilla.redhat.com/1428415
Bug-Url: https://bugzilla.redhat.com/1427782
Signed-off-by: Nir Soffer <nsoffer@redhat.com>