net: Introduce netupgrade module, processing the config upgrade
netupgrade is to be processed as an oneshot step at boot or during a
VDSM upgrade in order to reformat the networking configuration to a
supported format.

The step is exposed through vdsm-tool "upgrade-networks" command.

In this patch, the normalization steps once performed at kernelconf are
now moved here: "defaultRoute" and IFCFG keys.

Future patches will move to the netupgrade step the canonicalization of
the configuration and the creation of the unified configuration when
missing.

Change-Id: I4c6fa7e0f2679715c5ae0e0acd29111520b34983
Signed-off-by: Edward Haas <edwardh@redhat.com>