block-volume: Refactor shrink to optimal size logic
Currently, shrinkToOptimalSize performs the following:

1. Prepares the volume
2. Calls qemuimg-check to get volume actual size
3. Tears down the volume
4. Calls volume.reduce

The steps above will not work in cases where the volume is prepared and
we don't want to tear it down. In this patch, we introduce optimal_size
helper to return volume optimal size. With this helper,
shrinkToOptimalSize helper becomes redundant and thus it is removed.

Note: the optimal_size calculation is different that the
calculation performed in the removed shrinkToOptimalSize helper. In
shrinkToOptimalSize, the volume was reduced to actual size plus 0.5-1.5
GB depends on the volume size. In optimal_sze, the volume is
reduced to actual size plus a chunk.

This patch by mistake calls vol.UUID instead of vol.volUUID. This is
fixed in commit ae96cb6 (shrink-volume: Use vol.volUUID)

Change-Id: I0ee9a97d829ee71218e2896b7290eade11db0e62
Bug-Utl: https://bugzilla.redhat.com/1420405
Signed-off-by: Ala Hino <ahino@redhat.com>