net: Introducing common switch validations.
Some validation procedures are switch agnostic and should
be done before the setup call is dispatched to the relevant
switch.

An example for this is that regardless of switch type,
duplication of networks on a single southbound device
shouldn't occur -- "_validate_inter_network_compatibility"
goes to assure that, however it is currently only available
for legacy switches. The validation was moved up and is
called via _validate_common for each added network in
a single transaction.

Change-Id: I43f477f620dadd740dc2040f75971a30ee47d61b
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>