net: Create unified configuration from current host state
Create the persistent and running configs based on current host state
and maintain the configuration files to represent actual configuration.
The unified configuration is not to be consumed for ifcfg persitent
mode, but it is to be maintained (written to) from this point on.

Following this patch, changing the persistence mode from ifcfg to
unified should be straight forward.

Change-Id: I3de63e7bf82ce4eece7861f423ba3114b1f9a82a
Signed-off-by: Edward Haas <edwardh@redhat.com>