net: Modify unified config clear disk operation
Config.delete() has been implemented by deleting each network and bond
individually.
In order to allow config_exists() to represent if the configuration
exists or not and to simplify the delete steps, the networks and bonds
folders are removed with all their content in one shot.

Note: The save action will use a softer version named clearConfigs,
which leaves the nets and bonds root folders intact.

Change-Id: I97ba8a1af95b14df6e194993c6823a89da161862
Signed-off-by: Edward Haas <edwardh@redhat.com>