hostdev: report mdev supported types
This is part of VDSM side of vGPU implementation. We have to report
data from mdev-capable devices. Unfortunately, the kernel interface to
report concrete information regarding mdev type is insufficient at the
moment (the description sysfs file is pretty much vendor dependent
free-text), so we only report part of the information - mdev type name
and available instances. We also don't report driver at the moment as
we only support vfio-pci.

Change-Id: I24d5215abd5ff5db0e3982d2b6698d8630f6e367
Signed-off-by: Martin Polednik <mpolednik@redhat.com>