executor: bump the priority of logging
For executor, discarding workers it is unpleasant but definitely
possible. We should however not hide this, so we bump the messages
we emit when we discard one worker from DEBUG to INFO.

Having one executor exhausted is a different and much serious issue,
much less likely and much more serious. For this reason we change
the priority from DEBUG to WARNING.

Change-Id: Icfee3f313cdf7d43bb6b5fc3af4e0fd391d71516
Bug-Url: https://bugzilla.redhat.com/1428415
Backport-To: 4.1
Signed-off-by: Francesco Romani <fromani@redhat.com>