executor: store scheduled call as  attribute
We want to add periodic alert of blocked worker;
to do so, we start generalizing the action to be done on timeout.

The discard logic is  one-shot and this will not change.
However, the periodic alert will need to reschedule itself;
furthermore, when the worker stops we still need to cancel the
pending action.

To accomodate those needs, we will store the scheduled call
in one instance attribute.

Change-Id: I54386e940231585b00af5124e191a2d39d69cc10
Bug-Url: https://bugzilla.redhat.com/1428415
Backport-To: 4.1
Signed-off-by: Francesco Romani <fromani@redhat.com>