generate XML files for automatic driver installation
To get virtio support when installing Windows guest, user has to attach
floppy with virtio drivers and then manually load the drivers at the
beginning of the installation. By providing file Autounattend.xml with
paths to the drivers we can instruct windows to load the drivers
automatically and thus simplify the process for the user.

Unfortunately, due to various driver loading bugs in Windows, we cannot
have single XML file shared by all Windows version. The path to the
drivers must match only the drivers correct for the specific Windows
version being installed.

Change-Id: I784c4c83d5defcf425ff82b2d2b44288d7dcaae1
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>