network: add special treatment for dpdk ports
When nics are bound to DPDK driver, instead of traditional kernel
driver, they have to be accessed in a different manner.

Netowrk processing driver is a user space driver and packet processing
is not taking place in kernel domain.
Hence, traditional tools for reading nic speed, getting information
about its state and similar actions are irrelevant.

This patch adds fake information for nics which are bound to DPDK.

Change-Id: I9376dd3ddc45a931dc286cc05d4e57016532058b
Signed-off-by: Irit Goihman <igoihman@redhat.com>