net: Persist the bond slaves config in a sorted manner.
For consistency and to support the ability of comparing the requested
and current configuration, the bond slaves need to be persisted in a
sorted order.

This issue has been raised after the canonicalization step has been
moved from the config read to the netupgrade step (which normalizes the
config on boot and expects the application to maintain it as such from
that point forward).

Change-Id: I92d848e266a005400f62ef176fff76a152871ec3
Signed-off-by: Edward Haas <edwardh@redhat.com>