mkimage: make sure proper owner/group is set on payload directory
A minor fix to the creation of payload directory. In RPM package the
directory has vdsm/qemu as owner/group, but when the directory is
deleted and later recreated by supervdsm, owner/group remained set to
root. It is unlikely that this could have caused any real issues as only
supervdsm writes into that directory at the moment.

Change-Id: I26b03c250954d3fbd89292d962eb3329a50bafe5
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>