containers: fetch image name from metadata
In commit 46ba88dc557ee53adb8b3d20709a02a678910f6b
we started to remove smartness from the container domain parser.
The original plan was to get the image from the outside;
With the current plans to overhaul the domain metadata, it's
just simpler to add a new helper which parses it from the metadata.

This patch effectively restores part of the code removed in 46ba88dc55,
but now we call this explicitely, so we complete the removal
of unwanted and implicit smartness.

Change-Id: I78946d0ed8328d3196d8df1c461313258b5e25f0
Signed-off-by: Francesco Romani <fromani@redhat.com>