configfile: comment out old configuration when attribute is set by vdsm
When vdsm configures a key that already exists in conf file, the new key
will be created in vdsm-config-section. oldentries is a variable that tried to
do the opposite and leave keys instead of overriding with vdsm content if the
attribute is already set - the intention should be to use vdsm configuration
after calling vdsm-tool configure. This patch comment-out fields that
vdsm adds with a comment above those comments.

Change-Id: Id1dba75449b712432e0c2bc5b128cc996ede0153
Related-To: https://bugzilla.redhat.com/917062
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>