virt net: Create libvirt network only when needed
The libvirt network is needed only for VM/s with graphics, therefore its
creation and deletion should be dependent directly on it.

On each graphic entity setup, the network will be created if it has not
existed before.
Deletion will occur on teardown if the current instance is the last
one which uses the specific network.

This patch is focused on virt changes, the following patches drops
libvirt creation and deletion from the host networking code.

Change-Id: Ie55b08e9bdffcd8ea663b647c6d6dcf3363b9847
Signed-off-by: Edward Haas <edwardh@redhat.com>