virt: Add artificial delay after migration on destination
When a migration successfully finishes, we report two different events to
Engine:

- VM status change to up status from the destination.
- Final migration progress with downtime from the source.

Those events may arrive to Engine in arbitrary order.  But Engine developers
would like to already know the downtime at the time the migration is finished
from their view (i.e. they are informed that VM status is changed to up status
on the destination).

One solution would be to synchronize the two hosts before setting the VM to up
status on the destination.  Although that is possible, a simpler solution is
preferable.  The easiest workaround is to add some fixed delay after migration
is finished on the destination and before up status is set.  That doesn't
guarantee the proper order of events, but increases chances that the downtime
event is received by Engine before the up state event.  If the order of the
events is still wrong, downtime won't be displayed in Engine.  This is not much
problem if it happens only rarely.

Change-Id: I2b9500a0d9c7824a8b253817634a9a73eed23d27
Bug-Url: https://bugzilla.redhat.com/1427566
Backport-To: 4.1
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>