virt: periodic: HostMonitor: set not-discardable
Since we made HostMonitor one exclusive operation in change
I60ccd4e0e239ce5dfa2c90947bd6cd59a23e51b3
for the same reasons it also makes sense to set the `discard` flag
to False, so we have periodic reminder that the operation is stuck.

It works like this: with discard=False we implicitely schedule one
check after `timeout` seconds. If this fires, and if the worker is
stuck, one warning is emitted and the check is scheduled again
after `timeout` seconds.
This is indeed the behaviour we want there.

Please note that this change is a partial revert.
We need this because the executor 'periodic check,
not discard' feature was added while after the change
I60ccd4e0e239ce5dfa2c90947bd6cd59a23e51b3

Change-Id: If9d6923ad397c85909067747f0d7dd2a94dce8a1
Related-To: https://bugzilla.redhat.com/1419856
Backport-To: 4.1
Backport-To: 4.0
Signed-off-by: Francesco Romani <fromani@redhat.com>