sp: _upgradePool - msd update when needed
_upgradePool is executed during startSpm and a master domain update is
always initiated during its execution. Other operations (like storage
jobs such as copy_data/amend_image) that use the same lock may be running
which will cause the _upgradePool/startSpm operation to wait till those
operation end. In case of very long storage operation, that time may be
very long.

Usually no upgrade is actually being performed, this patch optimizes the
flow by verifying that a master domain upgrade is actually needed before
attempting to execute it (The domain version can be only incremented).
In the following patches I do the same for non master data domains.

Note that this is an optimization, there may be a case in which upgrade
will be needed and startSpm will wait for the lock, in that case the
running jobs should be aborted.

Change-Id: I93ef63b754209ba04555c5a99d90cd089a6eda18
Bug-Url: https://bugzilla.redhat.com/1430122
Signed-off-by: Liron Aravot <laravot@redhat.com>